Rafactor/repository

+++ backend/src/main/java/io/github/changmin6362/multistore/feature/authentication/api/AuthResponse.java
@@ -1,14 +1,17 @@
 package io.github.changmin6362.multistore.feature.authentication.api;
 
+import java.math.BigInteger;
+
 /**
  * 로그인/회원가입 성공 시 내려주는 응답
  * { userId, emailAddress, nickName, createdAt, accessToken, refreshToken }
  */
 public record AuthResponse(
-        Long userId,
+        BigInteger userId,
         String emailAddress,
         String nickName,
         String createdAt,
         String accessToken,
         String refreshToken
-) {}
+) {
+}
